+++ apps/models/form/Auth/UserLoginForm.php
@@ -51,13 +51,17 @@ class UserLoginForm extends Validator
 
     public static function inputRules()
     {
+        /**
+         * We only control frontend behaviour of input keys - `securelogin`, `logout`, `ssl`,
+         * So we not add Rules `Required` for these keys
+         */
         return [
-            'username' => 'required',
+            'username' => 'Required',
             'password' => [
-                ['required'],
-                ['length', ['min' => 6, 'max' => 40]]
+                ['Required'],
+                ['Length', ['min' => 6, 'max' => 40]]
             ],
-            'opt' => 'length(6)',
+            'opt' => ['Length', ['min' => 6, 'max' => 6]],
             'securelogin' => 'Equal(value=yes)',
             'logout' => 'Equal(value=yes)',
             'ssl' => 'Equal(value=yes)',